+++ backend/src/main/java/middle_point_search/backend/domains/timeVoteRoom/dto/dto/TimeRange.java
@@ -1,0 +1,29 @@
+package middle_point_search.backend.domains.timeVoteRoom.dto.dto;
+
+import java.time.LocalDateTime;
+
+import org.springframework.format.annotation.DateTimeFormat;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PUBLIC)
+public class TimeRange {
+
+	@JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+	@DateTimeFormat(pattern = "yyyy-MM-dd HH:mm")
+	@Schema(type = "string", example = "2024-07-26 13:00")
+	private LocalDateTime memberAvailableStartTime;
+
+	@JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+	@DateTimeFormat(pattern = "yyyy-MM-dd HH:mm")
+	@Schema(type = "string", example = "2024-07-26 19:00")
+	private LocalDateTime memberAvailableEndTime;
+}
