+++ backend/src/main/java/middle_point_search/backend/domains/timeVoteRoom/dto/request/VoteRequest.java
@@ -1,0 +1,10 @@
+package middle_point_search.backend.domains.timeVoteRoom.dto.request;
+
+import java.util.List;
+
+import middle_point_search.backend.domains.timeVoteRoom.dto.dto.TimeRange;
+
+public record VoteRequest(
+	List<TimeRange> dateTime
+) {
+}
